refactor: delete useCarForm/useServiceForm aliases, glassmorphism Account page (#223 #251)
- refactor(#251): delete useCarForm.js and useServiceForm.js — both were
  single-line re-exports of useFormData with a different name; update the
  two callers (useCarHandlers, useServiceHandlers) to import useFormData
  directly from ../../../hooks/useFormData
- fix(#223): account.css — add .account-page container with max-width and
  padding; add glassmorphism styling (.table-container with dark
  background, blur, border, rounded corners) matching Dashboard and
  Appointments; wrap Account.jsx content in <div className="account-page">

Closes #223, closes #251
