--- client/src/pages/cars/hooks/useCarForm.js
@@ -1,1 +1,0 @@
-export { useFormData as useCarForm } from "../../../hooks/useFormData";
