+++ client/src/pages/cars/hooks/useCarHandlers.js
@@ -1,4 +1,4 @@
-import { useCarForm } from "./useCarForm";
+import { useFormData as useCarForm } from "../../../hooks/useFormData";
 import { useCarActions } from "./useCarActions";
 import { useCarsUIStore } from "../../../stores/uiStores";
