--- client/src/pages/servicesAdmin/hooks/useServiceForm.js
@@ -1,1 +1,0 @@
-export { useFormData as useServiceForm } from "../../../hooks/useFormData";
