+++ client/src/pages/servicesAdmin/hooks/useServiceHandlers.js
@@ -1,4 +1,4 @@
-import { useServiceForm } from "./useServiceForm";
+import { useFormData as useServiceForm } from "../../../hooks/useFormData";
 import { useServiceActions } from "./useServiceActions";
 import { useServicesUIStore } from "../../../stores/uiStores";
